Rewrote to use Jackson annotations, semantics; now reproduces unexpected exception
